Create BackwardsCompatibleReadsRecipes
This contains all existing implicit HttpReads, leaving the traits as pure
sets of helper functions for the dsl.
