+++ src/main/scala/uk/gov/hmrc/play/http/reads/HttpReads.scala
@@ -21,11 +21,12 @@ import uk.gov.hmrc.play.http.HttpResponse
 trait HttpReads[O] {
   def read(method: String, url: String, response: HttpResponse): O
 }
-object HttpReads extends HtmlHttpReads with JsonHttpReads {
+object HttpReads extends BackwardsCompatibleReadsRecipes {
+
   // readRaw is brought in like this rather than in a trait as this gives it
   // compilation priority during implicit resolution. This means, unless
   // specified otherwise a verb call will return a plain HttpResponse
-  implicit val readRaw: HttpReads[HttpResponse] = RawHttpReads.readRaw
+  implicit val readRaw: HttpReads[HttpResponse] = BackwardsCompatibleReadsRecipes.readRaw
 
   def apply[O](readF: (String, String, HttpResponse) => O): HttpReads[O] = new HttpReads[O] {
     def read(method: String, url: String, response: HttpResponse) = readF(method, url, response)
