+++ src/test/scala/uk/gov/hmrc/play/Examples.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.play
 
 import uk.gov.hmrc.play.audit.http.HeaderCarrier
-import uk.gov.hmrc.play.http.reads.HttpReads
+import uk.gov.hmrc.play.http.reads.{JsonHttpReads, OptionHttpReads, BackwardsCompatibleReadsRecipes, HttpReads}
 import scala.concurrent.ExecutionContext.Implicits.global
 
 object Examples {
@@ -71,21 +71,20 @@ object Examples {
     import play.twirl.api.Html
     http.GET[Html]("http://gov.uk/hmrc") // Returns a Play Html type
 
-    object ExampleWithOptionReadsAsMixin extends reads.OptionHttpReads {
-      http.GET[Option[Html]]("http://gov.uk/hmrc") // Returns a None, or a Play Html type
-      http.GET[Option[MyCaseClass]]("http://gov.uk/hmrc") // Returns None, or Some[MyCaseClass] de-serialised from JSON
-    }
 
-    {
-      import reads.OptionHttpReads._
-      http.GET[Option[Html]]("http://gov.uk/hmrc") // Returns a None, or a Play Html type
-      http.GET[Option[MyCaseClass]]("http://gov.uk/hmrc") // Returns None, or Some[MyCaseClass] de-serialised from JSON
-    }
-
-    {
-      import reads.OptionHttpReads.{noneOn, some}
+    object ExampleWithOptionReads extends OptionHttpReads {
       implicit val myReads: HttpReads[Option[Html]] = noneOn(status = 204) or some[Html]
       http.GET[Option[Html]]("http://gov.uk/hmrc") // Returns a None, or a Play Html type
+    }
+
+    {
+      import OptionHttpReads._
+      import JsonHttpReads._
+
+      implicit val readOptionalMyCaseClass: HttpReads[Option[MyCaseClass]] =
+        noneOn(204) or noneOn(404) or some(jsonBodyDeserialisedTo[MyCaseClass])
+
+      http.GET[Option[MyCaseClass]]("http://gov.uk/hmrc") // Returns None, or Some[MyCaseClass] de-serialised from JSON
     }
   }
 }
