+++ src/test/scala/uk/gov/hmrc/play/http/HttpPostSpec.scala
@@ -42,7 +42,8 @@ class HttpPostSpec extends WordSpecLike with Matchers with CommonHttpBehaviour {
       val testPOST = new StubbedHttpPost(Future.successful(response))
       testPOST.POST(url, testObject).futureValue shouldBe response
     }
-    "be able to return HTML responses" in new HtmlHttpReads {
+    "be able to return HTML responses" in {
+      import uk.gov.hmrc.play.http.reads.BackwardsCompatibleReadsRecipes.readToHtml
       val testPOST = new StubbedHttpPost(Future.successful(new DummyHttpResponse(testBody, 200)))
       testPOST.POST(url, testObject).futureValue should be (an [Html])
     }
@@ -61,7 +62,8 @@ class HttpPostSpec extends WordSpecLike with Matchers with CommonHttpBehaviour {
       val testPOST = new StubbedHttpPost(Future.successful(response))
       testPOST.POSTForm(url, Map()).futureValue shouldBe response
     }
-    "be able to return HTML responses" in new HtmlHttpReads {
+    "be able to return HTML responses" in {
+      import uk.gov.hmrc.play.http.reads.BackwardsCompatibleReadsRecipes.readToHtml
       val testPOST = new StubbedHttpPost(Future.successful(new DummyHttpResponse(testBody, 200)))
       testPOST.POSTForm(url, Map()).futureValue should be (an [Html])
     }
@@ -80,7 +82,8 @@ class HttpPostSpec extends WordSpecLike with Matchers with CommonHttpBehaviour {
       val testPOST = new StubbedHttpPost(Future.successful(response))
       testPOST.POSTString(url, testRequestBody).futureValue shouldBe response
     }
-    "be able to return HTML responses" in new HtmlHttpReads {
+    "be able to return HTML responses" in {
+      import uk.gov.hmrc.play.http.reads.BackwardsCompatibleReadsRecipes.readToHtml
       val testPOST = new StubbedHttpPost(Future.successful(new DummyHttpResponse(testBody, 200)))
       testPOST.POSTString(url, testRequestBody).futureValue should be (an [Html])
     }
@@ -99,7 +102,8 @@ class HttpPostSpec extends WordSpecLike with Matchers with CommonHttpBehaviour {
       val testPOST = new StubbedHttpPost(Future.successful(response))
       testPOST.POSTEmpty(url).futureValue shouldBe response
     }
-    "be able to return HTML responses" in new HtmlHttpReads {
+    "be able to return HTML responses" in {
+      import uk.gov.hmrc.play.http.reads.BackwardsCompatibleReadsRecipes.readToHtml
       val testPOST = new StubbedHttpPost(Future.successful(new DummyHttpResponse(testBody, 200)))
       testPOST.POSTEmpty(url).futureValue should be (an [Html])
     }
